+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 ### Features
 
 * add community rating feature to feedback system ([#268](https://github.com/Deepwerks/discord-bot/issues/268)) ([77f6ae9](https://github.com/Deepwerks/discord-bot/commit/77f6ae98c11c86e80f0991d901afd125fc17a631))
-* added redirect routes to TOS and privacy policy ([429f7b7](https://github.com/Deepwerks/discord-bot/commit/429f7b78fa1bcc622d1a8c50753b3b3733db6c99))
-* added tos and privacy policy links to help command response ([d8d42a4](https://github.com/Deepwerks/discord-bot/commit/d8d42a46a5162834944adebe5e018533f8140d0c))
+* created sequelize models for guild subscription monitoring ([1368168](https://github.com/Deepwerks/discord-bot/commit/13681681c7e4cdceff8124c2bda2e7d6893f1bec))
+* implemented a function to check guild chatbot usage ([df46e72](https://github.com/Deepwerks/discord-bot/commit/df46e729942f5e32ce404a0fd3fa525a4d4cd282))
+* implemented guild subscriptions ([ee0d688](https://github.com/Deepwerks/discord-bot/commit/ee0d68826a439bffa56304700c32f0211ab58285))